pydanticai_agent/pyproject.toml
[project]
name = "pydanticai-agent"
version = "0.1.0"
description = "pydanticai agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic-ai>=0.3.4",
    "python-dotenv>=1.1.1",
]
